add initial cmake for flatbuffers
