+++ onnxruntime/core/flatbuffers/flatbuffers_utils.h
@@ -1,0 +1,26 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include <string>
+
+namespace onnxruntime {
+namespace experimental {
+namespace utils {
+
+template <typename T>
+bool IsOrtFormatModel(const std::basic_string<T>& filename) {
+  auto len = filename.size();
+  return len > 4 &&
+         filename[len - 4] == '.' &&
+         std::tolower(filename[len - 3]) == 'o' &&
+         std::tolower(filename[len - 2]) == 'r' &&
+         std::tolower(filename[len - 1]) == 't';
+}
+
+bool IsOrtFormatModelBytes(const void* bytes, int num_bytes);
+
+}  // namespace utils
+}  // namespace experimental
+}  // namespace onnxruntime
